Refactoring the condition tree. Tests still WIP.
